[Refactor] 파일 이동 및 import 경로 수정
