+++ src/components/book/TicketInfoComponent.vue
