+++ src/router/index.js
@@ -2,9 +2,9 @@ import ExchangePage from '@/pages/ExchangePage.vue';
 import LoginComponent from '@/components/auth/LoginComponent.vue';
 import MainPage from '@/pages/MainPage.vue';
 import MyPage from '@/pages/Mypage.vue';
-import PurchaseSuccessComponent from '@/components/reservation/PurchaseSuccessComponent.vue';
-import SelectPriceComponent from '@/components/reservation/SelectPriceComponent.vue';
-import SelectSeatComponent from '@/components/reservation/SelectSeatComponent.vue';
+import PurchaseSuccessComponent from '@/components/book/PurchaseSuccessComponent.vue';
+import SelectPriceComponent from '@/components/book/SelectPriceComponent.vue';
+import SelectSeatComponent from '@/components/book/SelectSeatComponent.vue';
 import { createRouter, createWebHistory } from 'vue-router';
 import { useMemberStore } from '@/stores/useMemberStore';
 import AgreeComponent from '@/components/auth/AgreeComponent.vue';
